requirements.txt
bidict==0.21.4
click==8.0.4
dnspython==2.2.1
eventlet==0.33.0
Flask==2.0.3
Flask-SocketIO==5.1.1
greenlet==1.1.2
importlib-metadata==4.11.2
imutils==0.5.4
install==1.3.5
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
numpy==1.21.5
opencv-python==4.5.5.62
python-engineio==4.3.1
python-socketio==5.5.2
PyYAML==6.0
random2==1.0.1
six==1.16.0
thread6==0.2.0
toml==0.10.2
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.7.0
